[FIX] web_responsive: Make showAppsMenuItem work again
In Odoo 17, the Class TourStepUtils was removed, and stepUtils is now an object instead of a class. When this module is installed, the tours stop working and throw the error:
Uncaught Error: Dependencies should be defined by an array: function (require) {.

This commit adapts the code to align with these changes, ensuring that tours work correctly.